Wip add tests

+++ tesTokenizer/NeTokenizer/Encoded.cs
@@ -3,7 +3,7 @@
     public class Encoded
     {
         public string[] Tokens { get; set; }
-        public int[] Ids { get; set; }
+        public long[] Ids { get; set; }
         public int[] Mask { get; set; }
         public int[] WordIds { get; set; }
     }
